+++ packages/toolbars-new/web/src/icons/H1Icon.tsx
@@ -14,7 +14,7 @@ const H1Icon = () => (
       <g
         id="Mobile-headings"
         transform="translate(-165.000000, -687.000000)"
-        fill="#000000"
+        fill="currentColor"
         fillRule="nonzero"
       >
         <path